LMS--1/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
textblob==0.17.1
PyJWT==2.8.0
